+++ src/murphy/market_data/provider.py
@@ -19,8 +19,7 @@ class MarketDataProvider(Protocol):
     def fetch_option_chain_snapshots(
         self,
         tickers: list[str],
-        min_dte: int,
-        max_dte: int,
+        min_dte: float,
+        max_dte: float,
         as_of: datetime | None = None,
     ) -> list[OptionSnapshot]: ...
-
